requirements.txt
Flask==2.0.3
Flask_Cors==3.0.10
pandas==1.4.1
python_dateutil==2.8.2
requests==2.27.1
